pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "tft-analytics"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "google-cloud-bigquery>=3.11.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.2.0",
]

[tool.setuptools.packages.find]
include = ["tft_analytics*"]
